MinorProject/expense_tracker_backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.1.1
cryptography==46.0.3
cycler==0.12.1
decorator==4.4.2
Django==4.2.24
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dj-database-url==3.0.1
docutils==0.20.1
filetype==1.2.0
fonttools==4.56.0
gunicorn==23.0.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.5
joblib==1.4.2
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
kiwisolver==1.4.7
matplotlib==3.7.5
moviepy==1.0.3
mysql-connector==2.2.9
mysql-connector-python==9.0.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
Pillow==9.2.0
proglog==0.1.12
pycparser==2.23
Pygments==2.19.2
PyJWT==2.9.0
PyMySQL==1.1.2
pyparsing==3.1.4
pypiwin32==223
python-dateutil==2.9.0.post0
pytube==15.0.0
pytz==2024.2
pywin32==311
requests==2.32.4
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
